# Conflicts:
#	jsapp/js/components/library/assetBreadcrumbs.scss

+++ kpi/deployment_backends/mock_backend.py
@@ -1,14 +1,24 @@
 # coding: utf-8
 import copy
+import os
+import re
 import time
 import uuid
 from datetime import datetime
+from typing import Optional, Union
+from xml.etree import ElementTree as ET
+try:
+    from zoneinfo import ZoneInfo
+except ImportError:
+    from backports.zoneinfo import ZoneInfo
 
-import pytz
 from deepmerge import always_merger
-from dicttoxml import dicttoxml
+from dict2xml import dict2xml
 from django.conf import settings
 from django.urls import reverse
+from django.utils import timezone
+from django.utils.translation import gettext as t
+from lxml import etree
 from rest_framework import status
 
 from kpi.constants import (
@@ -17,12 +27,20 @@ from kpi.constants import (
     PERM_CHANGE_SUBMISSIONS,
     PERM_DELETE_SUBMISSIONS,
     PERM_VALIDATE_SUBMISSIONS,
-    PERM_VIEW_SUBMISSIONS,
+)
+from kpi.exceptions import (
+    AttachmentNotFoundException,
+    InvalidXPathException,
+    SubmissionNotFoundException,
+    XPathNotFoundException,
 )
 from kpi.interfaces.sync_backend_media import SyncBackendMediaInterface
 from kpi.models.asset_file import AssetFile
-from kpi.utils.mongo_helper import MongoHelper
+from kpi.tests.utils.mock import MockAttachment
+from kpi.utils.mongo_helper import MongoHelper, drop_mock_only
+from kpi.utils.xml import edit_submission_xml
 from .base_backend import BaseDeploymentBackend
+from ..exceptions import KobocatBulkUpdateSubmissionsClientException
 
 
 class MockDeploymentBackend(BaseDeploymentBackend):
@@ -30,51 +48,83 @@ class MockDeploymentBackend(BaseDeploymentBackend):
     Only used for unit testing and interface testing.
     """
 
+    PROTECTED_XML_FIELDS = [
+        '__version__',
+        'formhub',
+        'meta',
+    ]
+
+    @property
+    def attachment_storage_bytes(self):
+        submissions = self.get_submissions(self.asset.owner)
+        storage_bytes = 0
+        for submission in submissions:
+            attachments = self.get_attachment_objects_from_dict(submission)
+            storage_bytes += sum([attachment.media_file_size for attachment in attachments])
+        return storage_bytes
+
     def bulk_assign_mapped_perms(self):
         pass
 
     def bulk_update_submissions(
         self, data: dict, user: 'auth.User'
     ) -> dict:
-
-        submission_ids = self.validate_write_access_with_partial_perms(
+        submission_ids = self.validate_access_with_partial_perms(
             user=user,
             perm=PERM_CHANGE_SUBMISSIONS,
             submission_ids=data['submission_ids'],
             query=data['query'],
         )
 
-        if not submission_ids:
+        if submission_ids:
+            data['query'] = {}
+        else:
             submission_ids = data['submission_ids']
 
         submissions = self.get_submissions(
             user=user,
-            format_type=SUBMISSION_FORMAT_TYPE_JSON,
-            submission_ids=submission_ids
+            format_type=SUBMISSION_FORMAT_TYPE_XML,
+            submission_ids=submission_ids,
+            query=data['query'],
         )
 
-        submission_ids = [int(id_) for id_ in submission_ids]
+        if not self.current_submission_count:
+            raise KobocatBulkUpdateSubmissionsClientException(
+                detail=t('No submissions match the given `submission_ids`')
+            )
 
-        responses = []
+        update_data = self.__prepare_bulk_update_data(data['data'])
+        kc_responses = []
         for submission in submissions:
-            if submission['_id'] in submission_ids:
-                _uuid = uuid.uuid4()
-                submission['meta/deprecatedID'] = submission['meta/instanceID']
-                submission['meta/instanceID'] = f'uuid:{_uuid}'
-                for k, v in data['data'].items():
-                    submission[k] = v
+            # Remove XML declaration from submission
+            submission = re.sub(r'(<\?.*\?>)', '', submission)
+            xml_parsed = etree.fromstring(submission)
 
-                # Mirror KobocatDeploymentBackend responses
-                responses.append(
-                    {
-                        'uuid': _uuid,
-                        'status_code': status.HTTP_201_CREATED,
-                        'message': 'Successful submission'
-                    }
-                )
+            _uuid, uuid_formatted = self.generate_new_instance_id()
 
-        self.mock_submissions(submissions)
-        return self.__prepare_bulk_update_response(responses)
+            instance_id = xml_parsed.find('meta/instanceID')
+            deprecated_id = xml_parsed.find('meta/deprecatedID')
+            deprecated_id_or_new = (
+                deprecated_id
+                if deprecated_id is not None
+                else etree.SubElement(xml_parsed.find('meta'), 'deprecatedID')
+            )
+            deprecated_id_or_new.text = instance_id.text
+            instance_id.text = uuid_formatted
+
+            for path, value in update_data.items():
+                edit_submission_xml(xml_parsed, path, value)
+
+            kc_responses.append(
+                {
+                    'uuid': _uuid,
+                    'status_code': status.HTTP_201_CREATED,
+                    'message': 'Successful submission',
+                    'updated_submission': etree.tostring(xml_parsed) # only for testing
+                }
+            )
+
+        return self.__prepare_bulk_update_response(kc_responses)
 
     def calculated_submission_count(self, user: 'auth.User', **kwargs) -> int:
         params = self.validate_submission_list_params(user,
@@ -94,11 +144,19 @@ class MockDeploymentBackend(BaseDeploymentBackend):
             }
         })
 
+    @property
+    def current_month_submission_count(self):
+        monthly_counter = len(
+            self.get_submissions(self.asset.owner)
+        )
+        return monthly_counter
+
+    @drop_mock_only
     def delete_submission(self, submission_id: int, user: 'auth.User') -> dict:
         """
         Delete a submission
         """
-        self.validate_write_access_with_partial_perms(
+        self.validate_access_with_partial_perms(
             user=user,
             perm=PERM_DELETE_SUBMISSIONS,
             submission_ids=[submission_id],
@@ -113,7 +171,6 @@ class MockDeploymentBackend(BaseDeploymentBackend):
                 }
             }
 
-        self._check_mock_mongo()
         settings.MONGO_DB.instances.delete_one({'_id': submission_id})
 
         return {
@@ -132,7 +189,7 @@ class MockDeploymentBackend(BaseDeploymentBackend):
              or
              {"query": {"Question": "response"}
         """
-        submission_ids = self.validate_write_access_with_partial_perms(
+        submission_ids = self.validate_access_with_partial_perms(
             user=user,
             perm=PERM_DELETE_SUBMISSIONS,
             submission_ids=data['submission_ids'],
@@ -167,7 +224,7 @@ class MockDeploymentBackend(BaseDeploymentBackend):
 
         return {
             'content_type': 'application/json',
-            'status': status.HTTP_204_NO_CONTENT,
+            'status': status.HTTP_200_OK,
         }
 
     def duplicate_submission(
@@ -176,16 +233,26 @@ class MockDeploymentBackend(BaseDeploymentBackend):
         # TODO: Make this operate on XML somehow and reuse code from
         # KobocatDeploymentBackend, to catch issues like #3054
 
-        self.validate_write_access_with_partial_perms(
+        self.validate_access_with_partial_perms(
             user=user,
             perm=PERM_CHANGE_SUBMISSIONS,
             submission_ids=[submission_id],
         )
 
-        duplicated_submission = copy.deepcopy(
-            self.get_submission(submission_id, user=user)
+        submission = self.get_submission(submission_id, user=user)
+        _attachments = submission.get('_attachments', [])
+        dup_att = []
+        if _attachments:
+            # not exactly emulating database id incrementing but probably good
+            # enough for the mock tests
+            max_attachment_id = max(a['id'] for a in _attachments)
+            for i, att in enumerate(_attachments, 1):
+                dup_att.append({**att, 'id': max_attachment_id + i})
+
+        duplicated_submission = copy.deepcopy(submission)
+        updated_time = datetime.now(tz=ZoneInfo('UTC')).isoformat(
+            'T', 'milliseconds'
         )
-        updated_time = datetime.now(tz=pytz.UTC).isoformat('T', 'milliseconds')
         next_id = max((
             sub['_id']
             for sub in self.get_submissions(self.asset.owner, fields=['_id'])
@@ -194,46 +261,87 @@ class MockDeploymentBackend(BaseDeploymentBackend):
             '_id': next_id,
             'start': updated_time,
             'end': updated_time,
-            'meta/instanceID': f'uuid:{uuid.uuid4()}'
+            'meta/instanceID': f'uuid:{uuid.uuid4()}',
+            'meta/deprecatedID': submission['meta/instanceID'],
+            '_attachments': dup_att,
         })
-        
-        settings.MONGO_DB.instances.insert_one(duplicated_submission)
+
+        self.asset.deployment.mock_submissions([duplicated_submission])
         return duplicated_submission
+
+    def get_attachment(
+        self,
+        submission_id_or_uuid: Union[int, str],
+        user: 'auth.User',
+        attachment_id: Optional[int] = None,
+        xpath: Optional[str] = None,
+    ) -> MockAttachment:
+
+        submission_json = None
+        # First try to get the json version of the submission.
+        # It helps to retrieve the id if `submission_id_or_uuid` is a `UUIDv4`
+        try:
+            submission_id_or_uuid = int(submission_id_or_uuid)
+        except ValueError:
+            submissions = self.get_submissions(
+                user,
+                format_type=SUBMISSION_FORMAT_TYPE_JSON,
+                query={'_uuid': submission_id_or_uuid},
+            )
+            if submissions:
+                submission_json = submissions[0]
+        else:
+            submission_json = self.get_submission(
+                submission_id_or_uuid, user, format_type=SUBMISSION_FORMAT_TYPE_JSON
+            )
+
+        if not submission_json:
+            raise SubmissionNotFoundException
+
+        submission_xml = self.get_submission(
+            submission_json['_id'], user, format_type=SUBMISSION_FORMAT_TYPE_XML
+        )
+
+        if xpath:
+            submission_tree = ET.ElementTree(
+                ET.fromstring(submission_xml)
+            )
+            try:
+                element = submission_tree.find(xpath)
+            except KeyError:
+                raise InvalidXPathException
+
+            try:
+                attachment_filename = element.text
+            except AttributeError:
+                raise XPathNotFoundException
+
+        attachments = submission_json['_attachments']
+        for attachment in attachments:
+            filename = os.path.basename(attachment['filename'])
+
+            if xpath:
+                is_good_file = attachment_filename == filename
+            else:
+                is_good_file = int(attachment['id']) == int(attachment_id)
+
+            if is_good_file:
+                return MockAttachment(pk=attachment_id, **attachment)
+
+        raise AttachmentNotFoundException
+
+    def get_attachment_objects_from_dict(self, submission: dict) -> list:
+
+        if not submission.get('_attachments'):
+            return []
+        attachments = submission.get('_attachments')
+        return [
+            MockAttachment(pk=attachment['id'], **attachment)
+            for attachment in attachments
+        ]
 
     def get_data_download_links(self):
         return {}
-
-    def get_enketo_submission_url(
-        self,
-        submission_id: int,
-        user: 'auth.User',
-        params: dict = None,
-        action_: str = 'edit',
-    ) -> dict:
-        """
-        Gets URL of the submission in a format FE can understand
-        """
-        if action_ == 'edit':
-            partial_perm = PERM_CHANGE_SUBMISSIONS
-        elif action_ == 'view':
-            partial_perm = PERM_VIEW_SUBMISSIONS
-        else:
-            raise NotImplementedError(
-                "Only 'view' and 'edit' actions are currently supported"
-            )
-
-        submission_ids = self.validate_write_access_with_partial_perms(
-            user=user,
-            perm=partial_perm,
-            submission_ids=[submission_id],
-        )
-
-        return {
-            'content_type': 'application/json',
-            'data': {
-                'url': f'http://server.mock/enketo/{action_}/{submission_id}'
-            }
-        }
 
     def get_enketo_survey_links(self):
         # `self` is a demo Enketo form, but there's no guarantee it'll be
@@ -263,6 +371,7 @@ class MockDeploymentBackend(BaseDeploymentBackend):
         user: 'auth.User',
         format_type: str = SUBMISSION_FORMAT_TYPE_JSON,
         submission_ids: list = [],
+        request: Optional['rest_framework.request.Request'] = None,
         **mongo_query_params
     ) -> list:
         """
@@ -290,7 +399,7 @@ class MockDeploymentBackend(BaseDeploymentBackend):
             self.mongo_userform_id, **params)
 
         # Python-only attribute used by `kpi.views.v2.data.DataViewSet.list()`
-        self.current_submissions_count = total_count
+        self.current_submission_count = total_count
 
         submissions = [
             MongoHelper.to_readable_dict(submission)
@@ -301,11 +410,11 @@ class MockDeploymentBackend(BaseDeploymentBackend):
             return submissions
 
         return [
-            dicttoxml(
+            dict2xml(
                 self.__prepare_xml(submission),
-                attr_type=False,
-                custom_root=self.asset.uid,
-            ).decode()
+                wrap=self.asset.uid,
+                newlines=False,
+            )
             for submission in submissions
         ]
 
@@ -317,12 +426,15 @@ class MockDeploymentBackend(BaseDeploymentBackend):
             'data': submission.get('_validation_status')
         }
 
-    def mock_submissions(self, submissions: list):
+    @drop_mock_only
+    def mock_submissions(self, submissions: list, flush_db: bool = True):
         """
         Insert dummy submissions into deployment data
         """
-        self._check_mock_mongo()
-        settings.MONGO_DB.instances.drop()
+        if flush_db:
+            settings.MONGO_DB.instances.drop()
+        count = settings.MONGO_DB.instances.count_documents({})
+
         for idx, submission in enumerate(submissions):
             submission[MongoHelper.USERFORM_ID] = self.mongo_userform_id
             # Some data already provide `_id`. Use it if it is present.
@@ -330,7 +442,7 @@ class MockDeploymentBackend(BaseDeploymentBackend):
             # or others do not.
             # MockMongo will raise a DuplicateKey error
             if '_id' not in submission:
-                submission['_id'] = idx + 1
+                submission['_id'] = count + idx + 1
             settings.MONGO_DB.instances.insert_one(submission)
             # Do not add `MongoHelper.USERFORM_ID` to original `submissions`
             del submission[MongoHelper.USERFORM_ID]
@@ -390,7 +502,7 @@ class MockDeploymentBackend(BaseDeploymentBackend):
                               data: dict,
                               method: str) -> dict:
 
-        self.validate_write_access_with_partial_perms(
+        self.validate_access_with_partial_perms(
             user=user,
             perm=PERM_VALIDATE_SUBMISSIONS,
             submission_ids=[submission_id],
@@ -427,10 +539,10 @@ class MockDeploymentBackend(BaseDeploymentBackend):
         Examples:
             {"submission_ids": [1, 2, 3]}
             {"query":{"_validation_status.uid":"validation_status_not_approved"}
-        
+
         """
 
-        submission_ids = self.validate_write_access_with_partial_perms(
+        submission_ids = self.validate_access_with_partial_perms(
             user=user,
             perm=PERM_VALIDATE_SUBMISSIONS,
             submission_ids=data['submission_ids'],
@@ -451,7 +563,7 @@ class MockDeploymentBackend(BaseDeploymentBackend):
             fields=['_id'],
         )
 
-        submissions_count = 0
+        submission_count = 0
 
         for submission in submissions:
             if not data['validation_status.uid']:
@@ -467,15 +579,29 @@ class MockDeploymentBackend(BaseDeploymentBackend):
                 {'$set': {'_validation_status': validation_status}},
             )
 
-            submissions_count += 1
+            submission_count += 1
 
         return {
             'content_type': 'application/json',
             'status': status.HTTP_200_OK,
             'data': {
-                'detail': f'{submissions_count} submissions have been updated'
+                'detail': f'{submission_count} submissions have been updated'
             }
         }
+
+    @staticmethod
+    def generate_new_instance_id() -> (str, str):
+        """
+        Returns:
+            - Generated uuid
+            - Formatted uuid for OpenRosa xml
+        """
+        _uuid = str(uuid.uuid4())
+        return _uuid, f'uuid:{_uuid}'
+
+    @property
+    def submission_count(self):
+        return self.calculated_submission_count(self.asset.owner)
 
     @property
     def submission_list_url(self):
@@ -488,17 +614,36 @@ class MockDeploymentBackend(BaseDeploymentBackend):
         return reverse(view_name,
                        kwargs={'parent_lookup_asset': self.asset.uid})
 
+    @property
+    def submission_model(self):
+
+        class MockLoggerInstance:
+            @classmethod
+            def get_app_label_and_model_name(cls):
+                return 'mocklogger', 'instance'
+
+        return MockLoggerInstance
+
     def sync_media_files(self, file_type: str = AssetFile.FORM_MEDIA):
         queryset = self._get_metadata_queryset(file_type=file_type)
         for obj in queryset:
             assert issubclass(obj.__class__, SyncBackendMediaInterface)
 
-    @staticmethod
-    def _check_mock_mongo():
-        # Ensure we are using MockMongo before deleting data
-        mongo_db_driver__repr = repr(settings.MONGO_DB)
-        if 'mongomock' not in mongo_db_driver__repr:
-            raise Exception('Cannot run tests on a production database')
+    @classmethod
+    def __prepare_bulk_update_data(cls, updates: dict) -> dict:
+        """
+        Preparing the request payload for bulk updating of submissions
+        """
+        # Sanitizing the payload of potentially destructive keys
+        sanitized_updates = copy.deepcopy(updates)
+        for key in updates:
+            if (
+                key in cls.PROTECTED_XML_FIELDS
+                or '/' in key and key.split('/')[0] in cls.PROTECTED_XML_FIELDS
+            ):
+                sanitized_updates.pop(key)
+
+        return sanitized_updates
 
     @staticmethod
     def __prepare_bulk_update_response(kc_responses: list) -> dict:
